Basic translation using react-i18next
Basic example of using internationalization
in the modern WebUI, powered by the
react-i18next library[1]. This has been
applied to the main title of the Active
and Stage users pages.

For this example, the default language
(english) is explicitely set (i.e., not
inferred by the browser language) and two
files for each of the different languages
have been defined: english and spanish.

To check the i18n functionality, a
Selector (shown in the Masthead) has
been defined to manually change the
language of the page.

The steps followed to provide i18n
functionality are:
- Install the libraries:
   >> npm install i18next react-i18next
       i18next-browser-languagedetector
- Create the english and spanish strings
under `i18n/` and `en`/`es` folders
- Create the `i18n.ts` configuration file
- In the `index.ts` file, wrap the `App`
component with the `I18nextProvider` one
- Perform translations using the
`useTranslation` hook (`t`) in the `ActiveUsers`
and `StageUsers` components

[1]- https://react.i18next.com/
Signed-off-by: Carla Martinez <carlmart@redhat.com>