Basic translation using react-intl (aka. FormatJS)
Basic example of using internationalization
in the modern WebUI, powered by the
react-intl library (aka. FormatJS)[1].
This has been applied to the main title
of the Active and Stage users pages.

For this example, the default language
(english) is explicitely set (i.e., not
inferred by the browser language) and two
files for each of the different languages
have been defined: english and spanish.

To check the i18n functionality, a
Selector (shown in the Masthead) has
been defined to manually change the
language of the page.

The steps followed to provide i18n
functionality are:
- Install the libraries:
  - >> npm install react-intl
- Create the english and spanish strings
under `lang/` folder
- Create the `i18n.ts` file to retrieve
the different languages
- In the `App.ts` file, wrap the content
with the `IntlProvider` component
- Perform translations using the
`FormattedMessage` component in `ActiveUsers`
and `StageUsers` components

Signed-off-by: Carla Martinez <carlmart@redhat.com>