Deleted elements table should handle 'Conditions'
The `DeletedElementsTable` component displays the
about-to-delete elements. This is data-specific and
need to be specified. Thus, the `Condition` data type
should be added as this will be used in the Settings
page to remove Inclusive and Exclusive conditions.

Also, the column names should display the custom
name if this is specified by the component.

Signed-off-by: Carla Martinez <carlmart@redhat.com>